a lot of stuff
changed the character controller to the gml version so that a pause feature could be added when ever the popup is true. bookshelf loads the two random books need for puzzle 3 every time the player steps into the study, data persists into the puzzle room but resets if the player leaves. Player can interact with yellow button in puzzle room to check if solved correctly or not, only indications right now are the debugging ones. only thing left for puzzle3 is to decide on is what happens when player solves it.